Final version

+++ app.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const bodyParser = require('body-parser');
 const mongoose = require("mongoose");
+const _ = require("lodash");
 
 const app = express();
 app.set('view engine', 'ejs');
@@ -45,45 +46,59 @@ app.get("/", (req, res) => {
 
 
 app.get("/:customListName", (req, res) => {
-    const customListName = req.params.customListName;
+    const customListName = _.capitalize(req.params.customListName);
 
     List.findOne({name: customListName}, (err, foundList) => {
         if (err) { console.log(err) }
         else {
-            if (foundList) {
-                console.log("Exits!");
+            if (!foundList) {
+                const list = new List({ name: customListName, items: defaultItems });
+                list.save();
+                res.redirect("/" + customListName)
             } else {
-                console.log("Doesn't exits!");
+                res.render("list", { listTitle: foundList.name, items: foundList.items });
             }
         }
     })
-
-    // const list = new List({
-    //     name: customListName,
-    //     items: defaultItems
-    // });
-
-    // list.save();
-    // res.render('list', { listTitle: 'Pendientes de trabajo', items: workItems });
 });
 
 
 app.post("/", (req, res) => {
     const itemName = req.body.newItem;
+    const listName = req.body.list;
+    
     const item = new Item({ name: itemName });
-    item.save();
-    res.redirect("/");
+
+    if (listName === "Today") {
+        item.save();
+        res.redirect("/");
+    } else {
+        List.findOne({name: listName}, (err, foundList) => {
+            if (err) { console.log(err.message) }
+            else {
+                foundList.items.push(item);
+                foundList.save();
+                res.redirect("/" + listName)
+            }
+        });
+    }
 });
 
 app.post("/delete", (req, res) => {
     const checkedItemId = req.body.checkbox;
-
-    Item.findByIdAndRemove(checkedItemId, err => {
-        if (err) { console.log(err) }
-        else { console.log(`Successfully deleted item ${checkedItemId}`) }
-    });
-
-    res.redirect("/")
+    const listName = req.body.listName;
+    
+    if (listName === "Today") {
+        Item.findByIdAndRemove(checkedItemId, err => {
+            if (err) { console.log(err) }
+            else { console.log(`Successfully deleted item ${checkedItemId}`) }
+        });
+        res.redirect("/")
+    } else {
+        List.findOneAndUpdate({name: listName}, {$pull: {items: {_id: checkedItemId}}}, (err, foundList) => {
+            if (!err) { res.redirect("/" + listName) }
+        });
+    }
 });
